+++ scribus/pdflib.cpp
@@ -3313,7 +3313,7 @@ void PDFlib::setTextCh(PageItem *ite, uint PNr, uint d, QString &tmp, QString &t
 		QString out2;
 		//CB Section numbering
 		//out2 = out.arg(PNr+doc.FirstPnum, -zae);
-		out2=out.arg(doc.getSectionPageNumberForPageIndex(PNr), -zae);
+		out2=out.arg(doc.getSectionPageNumberForPageIndex(PNr), -(int) zae);
 		chx = out2.mid(d-za2, 1);
 	}
 	uint cc = chx[0].unicode();
